+++ app/pages/_app.tsx
@@ -74,7 +74,9 @@ export default function App({
       <Head>
         <title key="title">visualize.admin.ch</title>
         <meta property="og:type" content="website" />
-        <meta property="og:title" content={"visualize.admin.ch"} />
+        <meta property="og:title" content={titleByLocale[locale]} />
+        <meta property="og:description" content={descriptionByLocale[locale]} />
+        <meta property="og:image" content={`${PUBLIC_URL}/og-image.png`} />
         <meta property="og:url" content={`${PUBLIC_URL}${asPath}`} />
         {federalTheme.preloadFonts?.map((src) => (
           <link
@@ -110,3 +112,17 @@ export default function App({
     </>
   );
 }
+
+const titleByLocale = {
+  de: "Visualisieren Sie Schweizer Open Government Data",
+  fr: "Visualisez les données ouvertes de l’administration publique suisse",
+  it: "Visualizza i dati aperti dell’amministrazione pubblica svizzera",
+  en: "Visualize Swiss Open Government Data",
+};
+
+const descriptionByLocale = {
+  de: "Erstellen Sie Visualisierungen von Datensätzen, welche durch den Linked-Data Dienst (LINDAS) des Bundesarchivs bereitgestellt werden und betten Sie diese in Ihre Webseite ein.",
+  fr: "Créez et intégrez des visualisations à partir des jeux de données du service LINDAS (Linked Data).",
+  it: "Crea ed incorpora visualizzazioni partendo dai dataset forniti dal servizio LINDAS (Linked Data).",
+  en: "Create and embed visualizations from any dataset provided by the LINDAS Linked Data Service.",
+};